FS#698 - handling of association test parameter in test case type

+++ org.faktorips.devtools.core/src/org/faktorips/devtools/core/ui/editors/testcasetype/NewRootParameterWizard.java
@@ -276,7 +276,7 @@ public class NewRootParameterWizard extends Wizard implements IBlockedValidation
      * Resets the state of the wizard, means clear all fields
      */
     public void resetWizard() {
-        if (newTestParameter != null){
+        if (newTestParameter != null && ! newTestParameter.isDeleted()){
             newTestParameter.delete();
         }
